fixes: 修复关联类型中文为空问题
