+++ src/scene_server/admin_server/imports.go
@@ -30,5 +30,5 @@ import (
 	_ "configcenter/src/scene_server/admin_server/upgrader/x18.11.19.01"
 	_ "configcenter/src/scene_server/admin_server/upgrader/x18.12.12.01"
 	_ "configcenter/src/scene_server/admin_server/upgrader/x18.12.12.02"
-	_ "configcenter/src/scene_server/admin_server/upgrader/x19.02.15.10"
+	_ "configcenter/src/scene_server/admin_server/upgrader/x18.12.12.03"
 )
