+++ src/scene_server/admin_server/upgrader/x18.12.12.03/fixAssociationTypeName.go
@@ -10,7 +10,7 @@
  * limitations under the License.
  */
 
-package x19_02_15_10
+package x18_12_12_03
 
 import (
 	"context"
