+++ src/scene_server/admin_server/upgrader/x18.12.12.03/pkg.go
@@ -9,32 +9,30 @@
  * either express or implied. See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-package x19_02_15_10
+package x18_12_12_03
 
 import (
 	"context"
 
-	"configcenter/src/common"
-	"configcenter/src/common/condition"
-	"configcenter/src/common/mapstr"
-	"configcenter/src/common/metadata"
+	"configcenter/src/common/blog"
 	"configcenter/src/scene_server/admin_server/upgrader"
 	"configcenter/src/storage/dal"
 )
 
-func fixEventSubscribeLastTime(ctx context.Context, db dal.RDB, conf *upgrader.Config) error {
-	cond := condition.CreateCondition()
-	cond.Field(common.BKOwnerIDField).Eq(common.BKDefaultOwnerID)
-	cond.Field(common.BKSubscriptionNameField).Like("process instance refresh")
+func init() {
+	upgrader.RegistUpgrader("x18.12.12.03", upgrade)
+}
 
-	data := mapstr.MapStr{
-		common.LastTimeField: metadata.Now(),
-	}
-
-	err := db.Table(common.BKTableNameSubscription).Update(ctx, cond.ToMapStr(), data)
+func upgrade(ctx context.Context, db dal.RDB, conf *upgrader.Config) (err error) {
+	err = fixAssociationTypeName(ctx, db, conf)
 	if err != nil {
+		blog.Errorf("[upgrade x18.12.12.03] fixAssociationTypeName error  %s", err.Error())
 		return err
 	}
-	return nil
+	err = fixEventSubscribeLastTime(ctx, db, conf)
+	if err != nil {
+		blog.Errorf("[upgrade x18.12.12.03] fixEventSubscribeLastTime error  %s", err.Error())
+		return err
+	}
+	return
 }
